Move test_lea into separate module test_machine_code_assembler.py
